IDEA-317185 Change toolbar insets according to the mockups
Two changes here:

- change the top/bottom insets of horizontal toolbars from 4 to 5;

- change the insets of vertical toolbars to be the same as
those of horizontal toolbars.

The second change is because some vertical toolbars only had correct
width because of a bug: their insets were applied before changing
the orientation to vertical. Now that the new JBInsets API fixes this bug, it turns out
that in fact the correct insets are the same for both vertical and horizontal.

Leave two methods for now as we may later want to change them independently.

GitOrigin-RevId: 4cc3eab6f991439626dfd8825998fdf00be90cd5