[terminal] IDEA-330183 Paint caret in the running command block
Caret API from the Editor is not suitable, because, for example, it moves the caret during selection, that is prohibited in terminal. Caret move and selection logic is highly connected, so there is no easy way to both restrict caret moves and make possible to select text.
Since, the caret in the running terminal is fully controlled by the shell and user unable to move the caret, the easier solution is to hide editor caret and paint our caret separately.

GitOrigin-RevId: 1e465d3ae5d77032201bef7a17d281532f4055be