IDEA-317185 Add JBInsets.update and being migrating from unwrap/unscale
The idea here is to make JBInsets able to update themselves.
This includes scaling and reloading updated values from the UI Defaults.
We introduce the key field for the UI Defaults, and use
the unscaledDefault field to store a hardcoded default in case the UI Defaults
don't have the required value.

To be able to fully utilize this new ability we need to start migrating
away from using constructors involving raw integer values and use
the copy constructor more, to preserve the key and the default value.

Another issue with using raw values is that the existing unscale()
method was using JBUI.unscale(), which relies on floating-point arithmetics
on scaled integer values, which seems like a bad idea in itself,
as it may yield incorrect results because of rounding errors.

Instead of relying on unscale/unwrap, we should use JBInsets more
and raw Insets less, so start by removing unwrap() calls from JBUI
and JBEmptyBorder and use JBInsets-first API instead.

To keep the unscale() for now, as it's used elsewhere in the codebase,
we rewrite it to use getUnscaled() if the insets value is a JBInsets instance
and just return the original value otherwise, as we assume that all non-JBInsets
values are already unscaled. If that's not the case somewhere, we need to fix
it there, instead of relying on hacks in JBInsets.

Use the new JBInsets.update() method in JBEmptyBorder for now,
and use UI Defaults keys in JBUI to create JBInsets instances using the new
capabilities.

GitOrigin-RevId: dc437c5a76b49c38a912af6af348694fc8ad8311